+++ src/main/java/nextstep/ladder/domain/person/People.java
@@ -1,0 +1,4 @@
+package nextstep.ladder.domain.person;
+
+public class People {
+}
