+++ src/main/java/nextstep/ladder/domain/person/Person.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.person;
 
 public class Person {
